refactor: changed module name for clarity
