+++ app/ldap_protocol/dns/clients/abstract_client.py
